requirements.txt
groq
python-dotenv
httpx
Flask
requests